backend/api/views.py
from rest_framework.response import Response
from rest_framework.decorators import api_view, permission_classes, parser_classes
from rest_framework.permissions import AllowAny

@api_view(['GET'])
@permission_classes([AllowAny])
def home(request):
    return Response({"data": "Hello World"})

@api_view(['GET'])
@permission_classes([AllowAny])
def optimize_prompt(request):
    return Response({"data": "Hello World"})
