backend/requirements.txt
aiohappyeyeballs
aiohttp
aiolimiter
aiosignal
annotated-types
anyio
asgiref==3.8.1
attrs
backoff
certifi==2024.8.30
cffi
charset-normalizer
cryptography
dataclasses-json
distro
dj-rest-auth
Django
django-allauth
django-cors-headers
djangorestframework==3.15.2
djangorestframework-simplejwt
frozenlist
grpcio
grpcio-tools
h11
h2
hpack
httpcore
httpx
hyperframe
idna==3.10
jiter
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
marshmallow
mem0ai
monotonic
multidict
mypy-extensions
neo4j
numpy
openai
orjson
packaging
pinecone
pinecone-client
pinecone-notebooks
pinecone-plugin-inference
pinecone-plugin-interface
portalocker
posthog
protobuf
pycparser
pydantic
pydantic-settings
pydantic_core
PyJWT
pypdf
python-dateutil
python-dotenv==1.0.1
pytz
PyYAML
qdrant-client
rank-bm25
regex
requests==2.32.3
setuptools
six
sniffio
SQLAlchemy
sqlparse
tenacity
tiktoken
tqdm
typing-inspect
typing_extensions
urllib3==2.2.3
voyageai
yarl
